+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 + Added a clickable area in the villager GUI to show the recipes.
 + All the careers are shown in the JEI Item collection in the right as spawn eggs (The don't spawn any entity).
 + Added a clickable are in the JEI recipe to show all the recipes of one career. (Click in the text: `Career: X`)
++ Added a CompactMode to reduce the amount of text in the screen